refactor: rename filter components with feature-specific naming convention
